DB Quality, Schnittstelle Delete Material
In Quality Viscosity von int auf decimal geaendert.
Schnittstelle material/materialId zur Abfrage der Qualitaetswerte umbenannt zu quality/materialId
Schnittstelle um MAterial zu loeschen

+++ MAWI_Context/Models/QualityModel.cs
@@ -12,7 +12,7 @@ namespace MAWI_Context.Models
         public int MaterialId { get; set; }
         public Nullable<int> Whiteness { get; set; }
         public Nullable<decimal> Absorbency { get; set; }
-        public Nullable<int> Viscosity { get; set; }
+        public Nullable<decimal> Viscosity { get; set; }
         public Nullable<int> Ppml { get; set; }
         public Nullable<decimal> DeltaE { get; set; }
